Refactor: Seperate type and default values for productEdit type
